# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Platform API is a RESTful API for managing ...",
        "title": "Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eng. Gabriel Coelho | InovaData development team",
            "url": "https://solude.tech",
            "email": "suporte@solude.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.1"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/user-service-logs": {
            "get": {
                "description": "Gets all user-service log entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserServiceLog"
                ],
                "summary": "Fetch all UserServiceLogs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.PublicUserServiceLog"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user-service-logs/{identifier}": {
            "get": {
                "description": "Gets a user-service log entry by a flexible identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserServiceLog"
                ],
                "summary": "Get a UserServiceLog by ID or user:XXX or service:XXX",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifier",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PublicUserServiceLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user-service-logs/{logID}": {
            "delete": {
                "description": "Removes a user-service log record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserServiceLog"
                ],
                "summary": "Delete a UserServiceLog",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserServiceLog ID",
                        "name": "logID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.PublicUserServiceLog": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}